requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
python-dateutil>=2.8.0
validators>=0.20.0
python-dotenv==1.1.0
supabase==1.0.3